Removed intervalTimeout when globalTimeout present
In order to prevent possible memory leaks from timeago intervals not automatically stopping when the timeago element is removed from the page. As well as removing having a unique intervals running for every timeago element. There is now one timeoutevent per interval when you include the globaltimeout plugin.

What this means is if you have 3000 timeago elements on one page and they are all set to timeout at 1 second there will only be 1 timeout instead of 3000. It will also remove the timeout if all 3000 of those elements are removed from the page. If you set multiple timeout refresh rates it will make that many timeout timers. As in if 1500 of those you want to timeout every minute then there will be 2 timers, one that occurs every second, and one that occurs every minute.

I have yet to test this in the test suite.